Make a single plane version of Threshold plugin
To support virtual stacks best with threshold it would be good if it only
called data min and max for the current plane (unless asked to do whole set
of data).

This code is an implmentation of such an approach. It almost totally works.
However I am unable to update the existing dialog's min/max field bounds
ranges to match the current plane's min/max values. The ranges seem to only
be settable once at initialization time. So by default we do need a full
pass over the data.

Once we have a simple way to set these ranges more than once this plugin can
be puled off the bench and used.