Fixed display bug in getPlanePosition
In AbstractDataView, all the non-xy plane positions are checked against
the corresponding lengths of the underlying dataset. As the xy planes
are skipped in this positioning, the index is offset by 2 when checking
length.

However, the index into the dataset was not being offset by 2, thus -
for example - the 3rd dimension's position was capped at the first
dimension's length.

The dataset index is now also offset by 2.