Move the ImageJ launcher to its own repository
The new repository is:
https://github.com/imagej/imagej-launcher

The launcher's development cycle differs from that of ImageJ2 core, and
the launcher is also used by ImageJ1 and Fiji. It makes sense to version
it separately and hence have it reside in its own Git repository. Doing
this also makes the ImageJ2 build much less complex.