Update release process to reflect new structure
The structure of the project has been improved, with all dependencies
resolving to release versions, and the ImageJ launcher being split into
its own project. As such, the release steps have changed substantially;
releasing a new version of ImageJ is now much simpler.

We now lean on the SciJava release-version.sh script to release each
component, which each have their own schedule and (SemVer) versioning.
The idea is to make it as easy as possible to make a new release of any
given component, such that downstream components can start depending on
it very quickly.