ij-legacy: offer a more meaningful explanation instead of a LinkageError
When the ImageJ 1.x classes were already defined, we cannot patch
them anymore. Explain the situation better.

This is as much to help future users of ij-legacy as this developer, who
has now spent three times more than an hour to figure out what was going
on, with exactly the same resolution: ij.IJ was already defined.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>