ij-legacy tests: make sure that a main class is found
On Windows, currently the AppUtils get a chance to determine the main
class only after we ran the LegacyHeadlessTest. This test changes the
current thread's name, on which AppUtils unfortunately relies. So simply
change it back.

The shortcomings of scijava-common to handle that situation gracefully
rather than throwing a NullPointerException when getBaseDirectory() was
called will be fixed there.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>