ij-legacy tests (Windows): make sure that ij.IJ is *not* loaded early
We really need to make sure that ij.IJ is not yet loaded before trying
to patch in the headless hacks. Unfortunately, it seems that Windows is
not only challenged when you ask it to perform things that are
outright easy on POSIX platforms, it also seems that it insists to load
the complete shebang when you simply ask for ij.IJ.class.

To work around this Windows-only problem, use the class ij.CommandListener
(which has no reference to any ImageJ 1.x class that might need patching
in the legacy headless mode) instead, because, really, all we need at that
point is the location of the .jar file containing the ImageJ 1.x classes.

Noticed by judicial Jonny Jenkins.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>