Add a simple Java agent to identify when ImageJ 1.x classes are loaded
When ImageJ 1.x' classes are loaded before the ij-legacy patching had
a chance, we have a problem. The biggest problem is that we only have
a chance to detect that situation when the ij-legacy patching takes
place, so there is no way to know which code path loaded the unpatched
ImageJ 1.x classes.

To help with that, this commit adds a Java agent that reports whenever
classes in the "ij" package (or any of its nested packages) is loaded.

We will refine this agent in subsequent commits.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>