LegacyJavaAgent: add the 'init' option
In the mode triggered by the 'init' option, the agent will simply force
the code patching of the ImageJ 1.x classes no matter what. This is the
safest and easiest method to fix issues when the CodeHacker complains
that the ImageJ 1.x classes were loaded before they could be patched.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>