LegacyJavaAgent: in debug mode, rewrite the class
There is one problem with the legacy Java agent: it reports all the
problematic code paths, but it does not stop there.

We cannot simply throw an exception in the ClassFileTransformer
because -- as documented in ClassFileTransformer's JavaDoc -- an
exception thrown in the transform() method is treated as if null
was returned, i.e. the next ClassFileTransformer is given a chance.

This change is an attempt to rewrite the classes with a static class
initializer that throws the exception when the class is loaded.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>