LegacyJavaAgent: use Javassist instead
In an earlier attempt, we tried to write the class file ourselves.
The rationale: we would like to be able to run a minimal
ij-legacy-agent.jar to be able to debug ij-legacy itself and
therefore would like to keep dependencies to a minimum.

However, the earlier approach was too simplistic: it is not enough
to rewrite the class to have a class initializer that throws our
desired exception. If a static method such as IJ.log() is called,
Java will happily match the method signature *before* initializing
the class. So we would have to add *all* valid methods to our
rewritten class.

So let's give up on that track and use Javassist instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>