Updater (cmd-line): allow to force a shadowing upload
Consider two update sites, say, 'ImageJ' and 'Fiji' where 'Fiji' has a
higher priority than 'ImageJ' (i.e. comes later in the list of update
sites). When the 'ImageJ' update site removes files that 'Fiji' relies
on, the users will get a problem.

Therefore, we need a way to upload files to 'Fiji' that will go away
from 'ImageJ'.

This is the way:

	java -jar ij-updater-core.jar \
		upload --update-site Fiji --force-shadow <file.jar>...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>