Updater (cmd-line): use better status when force-shadowing
Earlier, when we wanted to force-shadow an up-to-date file, we set the
status to local-only before doing so. But it is better to pretend that
it is modified because the internal state of an up-to-date file is
more similar to the state of a modified one than it is to a local-only
one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>