DatasetService: add I/O convenience methods
These methods allow you to open, save and revert datasets.

They are extremely similar to similar methods currently provided by
IOService, as well as the SaveAsImage command in ij-commands. The reason
to centralize them here is in preparation for a generalization of
IOService act as a HandlerServer for arbitrary IOPlugins. After that
happens, DatasetService will be the compile-safe place to open and save
datasets, whereas IOService will be the general-purpose place to open
and save data of any type.