Use new DatasetService I/O methods
In preparation for removal of the Dataset-specific I/O methods from
IOService, we switch over to using the shiny new DatasetService I/O
methods designated for this purpose.

Most of the changes are very straightforward and have resulted in much
simpler code. Note that the changes to SaveAsImage are a bit involved
though; although its complexity was reduced in some places, a few hacks
remain until SCIFIO's ImgSaver becomes better at N-dimensional saving.