DatasetIOPlugin: make low priority
The problem seems to be that SCIFIO is too aggressive about what it
thinks it can handle; e.g., when trying to open a .txt file, the
DatasetIOPlugin gets selected.

To avoid this, we simply move DatasetIOPlugin to the bottom of the list.
Although this is a hack which ultimately we should not need, determining
whether an arbitrary file is in *any* supported image format is going to
be an expensive operation no matter what, so for performance reasons it
makes sense to leave the priority of this I/O plugin lower anyway.