LUTFileDragAndDropHandler: create new LUT display
At the moment, passing a ColorTable to the DisplayService's
createDisplay method does not work, but that will be remedied shortly.
So in anticipation of that, let's change the LUTFileDragAndDropHandler
to make use of it directly.

This eliminates the use of the LUTService's special-case createDisplay
method, in favor of the DisplayService's general-purpose method.