Avoid NullPointerException in DefaultImageDisplay
As pointed out by Tyler Corbin, there is a NullPointerException to be
expected when using the variable 'dataView' when it is still 'null'.

This is due to some left-over code from 03817156(DefaultImageDisplay:
add support for ColorTables): the code to make the dataView from data
(unless dataView is already available) was intended to be moved below,
not copied (the idea was obviously the DRY principle: Don't Repeat
Yourself -- ironic that I always repeat the explanation what DRY stands
for...).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>