ij-legacy: Port fiji-compat's patch to ij.macro.Functions#call()
Sometimes, changes are not accepted into ImageJ 1.x' classes even if
contributed by the person who contributed the code in question.

This is the case, for example, for the call() handling in the macro
interpreter: its handler for InvocationTargetExceptions still uses
non-DRY code rather than handing off to IJ.handleException().

Usually, repeated code is just ugly and hard to maintain. In this case,
it is also a bug: if IJ.setExceptionHandler() is used to install
a different exception handler than ImageJ 1.x' default one, the macro
interpreter will still insist on doing its own exception handling.

That does not show up in standard ImageJ 1.x workflows, of course,
because they rarely entail a call to IJ.setExceptionHandler().

As long as ImageJ 1.x carries that bug, we must carry this bug fix
in ij-legacy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>