Early work on runtime type discovery
This code is preliminary work on making ImageJ discover the data types it
can handle at runtime. The data types wrap Imglib types. The type converter
code will use this discovery and casting to change types. There are examples
of new types here too. This shows that extensibility is possible.