Add an Icon column to the update sites dialog
It looks for a favicon.png off the root of each site URL.

This initial implementation is a proof of concept, and delays the
appearance of the dialog by a noticeable margin (~1 sec on my system).
We could alternately construct the ImageIcons asynchronously or
implement some other means to ease the perceived pain here. My biggest
practical concern is what happens if an update site is down; would an
HTTP request timing out block the dialog from appearing during the
timeout period?