LegacyHeadless: avoid patching multiple times
When ImageJ 1.x is already patched for headless operation, there is
little reason to do it again. Besides, if we do it again, it will
lead to problems because of ClassFormatErrors.

LegacyHeadless#patch() can actually be called multiple times,
e.g. in our test suite when running LegacyHeadlessTest#testHeadless
(which forces headless patching) in a JRE whose GraphicsEnvironment
is headless (hence triggering headless patching again).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>