Fix DefaultUploaderService#installUploader()
We used initialize() to force a re-discovery of all available
uploaders after installing an uploader, but that code changed now:
initialize() no longer discovers all the uploaders; uploadMap()
does. And hasUploader() calls uploadMap(). So all we need to do now is
to make sure that the DefaultUploaderService forgets what uploaders
it knows after installing a new uploader so that the lazy code can
kick in.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>