Make About ImageJ more robust
We will have to revisit this a bit later because it is still too
fragile.

The 'About' images should not lie about in the file system where
users can easily and inadvertently make a mess of them.

Also, the fall-back image should not be an ugly and boring black
image ;-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>