Remove the 'list script engines' plugin
The javax.scripting infrastructure never really caught on, even if
we use parts of it in our scripting framework. However, we finagle
compatibility with the javax.scripting API into the scripting languages
we want to support, and only enough that we can use them, not enough
to make them proper script engines.

Therefore, it is unlikely that the 'list script engines' plugin would
ever list more than 'Mozilla Rhino'. Which is a good reason to just
get rid of it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>