MediaWikiClient: add default ctor w/ FIJI_WIKI_URL
This makes it more convenient to construct a MediaWikiClient
specifically for use with the ImageJ/Fiji wiki.

This change also adds a TODO to remove this Fiji-specific reference from
the ImageJ2 codebase. It might work to simply use wiki.imagej.net as is,
but I didn't test it, so I left it as a TODO for now.