Updater: immediately initialize & add all update sites
Previously, the sites were only discovered and added when the
SitesDialog was actually popped up. This caused the FilesCollection data
structure to be incomplete and incorrect until such a GUI element was
constructed and presented.

The new behavior properly and fully initializes the FilesCollection much
earlier, so that it can be interrogated as needed in various places.