ij-legacy: Do not throw exceptions when legacy patching failed
When starting Fiji with another ImageJ 1.x version than the one we
compiled against, it is not a good idea to fail patching with an
exception and prevent Fiji from starting.

Better to report the failure but still try to start up, in line with
Postel's Law: be conservative in what you do, be liberal in what you
accept from others.

However, when we run unit tests, we want to fail loudly when patches
failed...

This commit uses the CodeHacker#maybeThrow() infrastructure to throw
the exceptions in unit testing mode and to dump them to stderr and
try to continue otherwise.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>