Prevent the unit tests from opening a Java cup on MacOSX
On MacOSX, the unit tests which instantiate UI elements -- even if
they do not show them -- open a separate process with a separate dock
icon. This is pretty disruptive when developers run all the unit tests
in the background and try to do something else such as catching up
with their mail.

The apple.awt.UIElement property helps with that.

Suggested by David Philip Marshall in

	http://mid.gmane.org/1374227295612-5004030.post@n6.nabble.com

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>