ij-legacy: move MacAdapter handling from LegacyInjector to LegacyExtensions
The MacAdapter patches are really needed for the legacy mode, and we
decided some time ago that that category of legacy patches should be
separated from the patches synchronizing legacy and modern state.

Technically, LegacyInjector *would* be the correct location if we still
just switched off the MacAdapter. However, we cannot switch it off at
all times anymore because that would break legacy mode (and hence Fiji,
e.g. when the user wants to quit the application)...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>