ij-legacy: do not let MacAdapter do its thing
If we let MacAdapter run normally, we will not be able to install any
other application listener. But ImageJ2 needs to.

For the moment, this breaks Fiji's legacy mode again: neither Quit nor
Drag 'n Drop of files onto the Dock icon will work. We will fix that in
the next commits.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>