Convert the legacy integration tests to use the maven-failsafe-plugin
The maven-invoked-plugin is very powerful, too powerful in fact to
work efficiently with integration tests. For example, we cannot easily
run them in Eclipse, and we als cannot easily skip them with -DskipITs
(and imply that with -DskipTests) but have to jump through hoops.

Better to use the maven-failsafe-plugin which was designed for
integration tests.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>