populate-app: do not insist on removing the icon and tools.jar
The ImageJ launcher now creates a .desktop icon on Linux by default. We
do not really need to remove that to obtain a clean ImageJ.app/.

Likewise, the jars/tools.jar is still useful for users running in a
JRE (as opposed to a JDK), so let's not get rid of that, either.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>