Mark @Parameters private in UpdatesAvailable
The original idea was that we wanted to re-use this class in Fiji
where no context was available at the time.

In the meantime, there *is* an ImageJ2 context available, so we do not
need to play games with setting fields of commands by hand anymore,
i.e. there is no need for the fields to be public anymore.

Careful analysis of both fiji.git and imagej.git shows that this
plugin was never used in Fiji, so let's simply make those fields
private. No harm done.

Noticed by Barry DeZonia.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>