Widgets now refresh off the EDT
All UI-based widget implementations will now run their refreshWidget
methods off the EDT. This will allow those methods to perform dispatch
operations, such as sending status updates via the StatusService (which,
if launched from the EDT, would deadlock).

To accomplish this, added an AbstractUIInputWidget layer to mark widgets
that are UI-specific and spawn new threads if a given widget is a part
of a UI that uses the EDT.

he UserInterface interface was modified with a flag method indicating if
a UI implementation requires the EDT or not.

All UI implementations were updated as appropriate, and all
AbstractUIInputWidgets shifted their refreshWidget implementations to
doRefresh (the public refreshWidget is now the gateway method where
thread-safety is checked).