WidgetModel: fix getMin/Max javadoc
It was slightly incorrect before; the reality is that it only gives back
null if the type is unbounded (e.g., BigInteger). For any bounded type
(e.g., double), you will get back a minimum/maximum bound corresponding
to the absolute minimum/maximum of that type, even if no min/max
attribute was specified in the @Parameter annotation.