BrightnessContrast: set soft minimum & maximum
This makes the behavior more consistent with ImageJ1: the Minimum and
Maximum fields now have sliders which are bounded by the *data* minimum
and maximum. You can still use the spinners to explicitly set the
minimum and maximum values outside the range of the sliders (by either
explicitly typing a number or by using the spinner arrow buttons), but
if you use the sliders directly you will be bounded by the data min/max.