Updater: complete the Fiji Wiki account hint
When registering a new update site, we have a convenient way to register
a new Fiji Wiki account at the same time -- if one is not yet available
to the current user.

However, the helpful hint was cut off, or rather, never completed by
this developer.

Pointed out by Ulrik Günther.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>