ij-legacy: work around limitations of the DummyLegacyService
Even in the ImagePlus patches, we should not interact with the
LegacyImageMap as long as the DummyLegacyService is active.

This change was really intended to be part of its parent commit,
but this developer forgot that this incredibly well-written
software called Eclipse sometimes (in fact, quite often now)
ignores Ctrl+S and simply does not save the current buffer.

No cookie for you, Eclipse.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>