CommandModule: do a full context injection
This may ultimately not be the best way, but it fixes the critical
ImageJ runtime bug with commands. On the SciJava Common side, I will
add additional safeguards when performing context injection, and write
corresponding unit tests, to ensure the DI is working as required.

Still remaining to be fixed: SCIFIO-related bug (i.e., File > Open).