TestBinaryMaskOverlay: remove unneeded try/catch
In the same vein as 7c0ea1fda02d4e56793eabfaa5761a8965a0e903, it is
better for tests which are not supposed to throw exceptions to simply
declare those checked exceptions, such that if they *are* thrown, the
test is in error. Then we get the full stack trace instead of just the
message, too.