CommandInfo: remove factory infrastructure
This infrastructure was put in place long ago to accommodate a specific
use case: 3D Viewer plugins, which were special extensions of Commands
with their own menu structure separate from vanilla ImageJ commands.

That use case can now be handled using the menuRoot attribute of the
@Plugin annotation; there is no more need to involve factories that
customize the wrapping of Commands to Module instances.

To reduce the complexity of the class hierarchy surrounding Commands and
Modules (which is really more than complicated enough already), we
remove this unused feature.

If the need should arise for it in the future, it can always be readded
at that time. But until then, we respect the "YAGNI" principle here.