Add ModuleCommand abstract class
This is another alternative abstract base class for Command
implementations. It is a Command that implements Module directly,
similar to DynamicCommand, but with DynamicCommand's baggage of being a
MutableModule.

The first use case is for OptionsPlugin, which really does *not* need to
be a MutableModule, but *does* need to inspect and mutate its own inputs
and outputs programmatically in its load() and save() methods.

This will allow OptionsPlugins to be used more easily headlessly in
programs like CellProfiler and KNIME, which for practical reasons really
cannot allow DynamicCommands to execute since they do not easily fit
within the fixed workflow paradigms of those systems.

The name "ModuleCommand" is rather unfortunately similar to
"CommandModule" and hence I expect will cause confusion. But I really
could not come up with a more appropriate name. Suggestions welcome.