DefaultModuleService: reuse module instances
Specifically, if a module instance is registered with the object service
as the sole instance of a particular type, the createModule(ModuleInfo)
method is now smart enough to figure that out, returning the existing
registered instance rather than creating a new one.

This is particularly useful for modules which are also SingletonPlugins,
which is the plan for OptionsPlugins. In this way, an OptionsPlugin
instance will be repeatedly executable via the ModuleService using its
CommandInfo, without the mess of temporary OptionsPlugin instances.