Avoid loading the MacOSX platform prematurely
We need to make sure that the platform classes load, and even
instantiate, correctly before we can ask them whether they target the
current platform. Therefore, we need to avoid loading MacOSX-specific
classes, such as the com.eawt.* classes. Unfortunately, if we
enumerate the declared fields, and one of them is of a subclass of
a MacOSX-specific class, that class *will* be loaded.

To work around this, simply do not declare the ultimate type of the
MacOSX-specific event handler, and everything's groovy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>