ij-legacy: always apply the black list for legacy commands
We should not make the same mistake as Microsoft's Visual C++ compiler:
behave differently in debug vs non-debug mode, because that makes
debugging a real bummer.

This also has the side effect of accelerating the start-up time
substantially: we just shaved off another 150 milliseconds on this
developer's machine, which contributes to the speed-up for the final
goal of having a perceived quick start-up time of less than a second
for the ImageJ2 context.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>