Updater: Add an API to specify the updateable platforms
In scenarios such as keeping an update site for Micro-Manager
up-to-date, it is often not only convenient, but even required to be
able to set the list of platforms for which to upload/download. This
commit adds that functionality.

As a special convenience, specifying "all" as a platform adds all the
platforms the updater knows about.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>