Updater (upload-complete-site): handle platform-specific files carefully
We do not necessarily want to remove files from the update site we are
uploading to when said files are specific to platforms other than the
one we are uploading from (and hence not installed locally).

Let's skip removing all files we do not expect to have because of a
different platform.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>