Updater (cmd-line): hide the WebDAV password when uploading
For convenience, we allow the WebDAV password to be stored
in the 'upload host' part of the upload site information:
webdav:<username>:<password>. However, we also showed the complete
upload host line in the progress when uploading -- particularly
inconvenient when uploading from a Jenkins job whose console output
is visible to everybody.

Special-case webdav to hide the password, with the intention of
deprecating that feature in favor of a $HOME/.netrc based solution.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>