Rename TextFieldWidget to TextWidget
The name "TextFieldWidget" was inconsistent with all the other widgets:
none of the others hardcode assumptions about the exact sort of widget
into the interface name itself. E.g., ObjectWidget is not called
ObjectChooserWidget or ObjectDropdownWidget. Hence, it is better to use
the simpler name "TextWidget" which can be used for text fields but also
other possibilities such as text areas.