Split legacy plugin finder tests into two classes
The main reason for this is to avoid calling
DefaultLegacyService.preinit() in the static initializer, because when
you call it, it masks a bug with legacy plugin discovery when
java.awt.headless is set to true.

With this change, adding:

    System.setProperty("java.awt.headless", "true");

At the beginning of either class's test method results in test failure,
which will need to be debugged.