Update Axes usage
Bumps pom-scijava to version 1.81, to take advantage of the latest
SCIFIO and ImgLib2 Axes API.

Axes was refactored from an enum to a static utility class, and some of
the constants (unused at the scijava-common level) were migrated to
SCIFIO.

Also, with the change from an enum to a utility class, some of the API
changed slightly or was deprecated in preparation of a transition to a
full N-Dimensional metadata model. Thus, constant references and method
calls were updated as needed.