ij-legacy (headless): safe-guard against unknown GenericDialog methods
Every once in a while, GenericDialog acquires new functionality that is --
of course -- incompatible with the headless mode. Rather than catch these
problems at the most inconvenient moment, such as in the final stages of a
paper or grant submission, let's catch them when they occur.

To that end, we introduce a unit test whose sole purpose is to figure out
whether there are methods in whatever GenericDialog version we want to use
which the LegacyGenericDialog does not account for yet.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>