Updater: prepare the update site list parser to a different column order
We had a broken updater for a day due to heavy changes in the layout
of http://fiji.sc/List_of_update_sites. Partially, the changes made
sense, even if they broke the updater. By preparing the updater
for those changes, we can make those changes again, after the vast
majority will have a new version of the updater already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>