change behavior of DefaultAutoscaleMethod
It has been the case for a while that we would never return a 0 width
interval (i.e. when data min == data max). However the existing code would
fail for some edge cases and give life to annoying display bugs.

Now we've decided that if data min == data max we will return an interval of
[typeMin, typeMax].