RestructureUtils: improve createNewImgPlus method
This removes uses of the deprecated calibration approach in favor of
cloning the input CalibratedAxis objects properly.

This change also introduces a proper generically typed method for
creating a new ImgPlus from an input ImgPlus template. It retains the
Dataset signature but leans on the new method to do its work.