EditAxes: remove superfluous error message
A null dataset cannot come in via the command framework because by
default, each @Parameter is required=true, so the preprocessing chain
will barf with "A Dataset is required but none exists" or some such
before run() is actually called.

In the case of the command being called via the Java API, it is possible
that run() could be called without setDataset having been called first.
In that case, the dataset could be null. But you will see the problem
very quickly in any case because getAxes() will throw
NullPointerException right away trying to interrogate the dataset.

So I think the log message is not necessary.