Updater (cmd-line): optionally forget missing dependencies upon upload
This adds support for the --forget-missing-dependencies option when
uploading: It will force the updater to break the dependency whenever a
dependency goes missing.

This option is intended primarily for automated uploads, e.g. the
Bio-Formats daily site.

The reason for this change was that the Bio-Formats daily site could no
longer be updated properly when the netcdf dependency was renamed into
netcdf-java: the updater believed that bio-formats needed both and
complained that netcdf-java was no longer there.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>