AbstractUserInterface: don't show displays twice
If a display is already being shown, the show method just returns.

This may become a problem if we decide to support multiple simultaneous
UIs, but for now is better than the display popping up multiple times,
which was previously the case if you manually called
UIService.show(Object) followed by UIService.showUI().