DefaultUIService: remove warning if no viewer
We now use the getDisplayViewer method to determine whether a display is
already being shown in the UI. But this caused the warning to appear
every time a display was not currently being shown. It is not really an
invalid state for a display to have no viewer instance yet.