DefaultUIService: fix race condition in lazy init
I was lazy when converting services over to lazy initialization, in that
I did not handle multiple threads properly. It is not actually that hard
to do: just make a boolean state flag, synchronize the initialization
method, setting the state flag upon completion and returning early if
the flag is already tripped. In this way, if a bunch of threads all try
to call a service's methods at once and trigger multiple lazy
initialization calls, only one of them will actually do the work and the
others will block until initialization is complete.

This change fixes a race condition when launching ImageJ with the
"--run" argument.

We should replicate this pattern for the other lazily instantiated
services...